[Chore] #1 - 폴더구조 정리
