+++ Sopt38-Seminar/Sopt38-Seminar/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
